requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
requests
werkzeug
sqlalchemy
google.genai
